refactor(structure): reorganize directory structure by feature
