+++ client/src/features/Header/Header.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
-import EditButton from "@/components/atoms/headerAtoms/EditButton/EditButton";
-import FridgeSelector from "@/components/atoms/headerAtoms/FridgeSelector/FridgeSelector";
+import EditButton from "@/features/Header/atoms/EditButton/EditButton";
+import FridgeSelector from "@/features/Header/atoms/FridgeSelector/FridgeSelector";
 
 // ⚠️将来 Redux に切り出す予定の型
 export type SelectedView = "fridge" | "expired";
